+++ Libs/Swagger/Client/Model/GetCorporationsCorporationIdWalletsDivisionTransactions200Ok.php
@@ -5,7 +5,7 @@
  * PHP version 7
  *
  * @category Class
- * @package  WordPress\Plugin\EveOnlineIntelTool\Libs\Swagger\Client
+ * @package  Swagger\Client
  * @author   Swagger Codegen team
  * @link     https://github.com/swagger-api/swagger-codegen
  */
@@ -36,7 +36,7 @@ use \WordPress\Plugin\EveOnlineIntelTool\Libs\Swagger\Client\ObjectSerializer;
  *
  * @category Class
  * @description wallet transaction
- * @package  WordPress\Plugin\EveOnlineIntelTool\Libs\Swagger\Client
+ * @package  Swagger\Client
  * @author   Swagger Codegen team
  * @link     https://github.com/swagger-api/swagger-codegen
  */
@@ -61,7 +61,7 @@ class GetCorporationsCorporationIdWalletsDivisionTransactions200Ok implements Mo
         'date' => '\DateTime',
         'type_id' => 'int',
         'location_id' => 'int',
-        'unit_price' => 'float',
+        'unit_price' => 'double',
         'quantity' => 'int',
         'client_id' => 'int',
         'is_buy' => 'bool',
@@ -78,7 +78,7 @@ class GetCorporationsCorporationIdWalletsDivisionTransactions200Ok implements Mo
         'date' => 'date-time',
         'type_id' => 'int32',
         'location_id' => 'int64',
-        'unit_price' => 'float',
+        'unit_price' => 'double',
         'quantity' => 'int32',
         'client_id' => 'int32',
         'is_buy' => null,
@@ -406,7 +406,7 @@ class GetCorporationsCorporationIdWalletsDivisionTransactions200Ok implements Mo
     /**
      * Gets unit_price
      *
-     * @return float
+     * @return double
      */
     public function getUnitPrice()
     {
@@ -416,7 +416,7 @@ class GetCorporationsCorporationIdWalletsDivisionTransactions200Ok implements Mo
     /**
      * Sets unit_price
      *
-     * @param float $unit_price Amount paid per unit
+     * @param double $unit_price Amount paid per unit
      *
      * @return $this
      */
