+++ Libs/Swagger/Client/Model/GetMarketsRegionIdOrders200Ok.php
@@ -5,7 +5,7 @@
  * PHP version 7
  *
  * @category Class
- * @package  WordPress\Plugin\EveOnlineIntelTool\Libs\Swagger\Client
+ * @package  Swagger\Client
  * @author   Swagger Codegen team
  * @link     https://github.com/swagger-api/swagger-codegen
  */
@@ -36,7 +36,7 @@ use \WordPress\Plugin\EveOnlineIntelTool\Libs\Swagger\Client\ObjectSerializer;
  *
  * @category Class
  * @description 200 ok object
- * @package  WordPress\Plugin\EveOnlineIntelTool\Libs\Swagger\Client
+ * @package  Swagger\Client
  * @author   Swagger Codegen team
  * @link     https://github.com/swagger-api/swagger-codegen
  */
@@ -63,7 +63,7 @@ class GetMarketsRegionIdOrders200Ok implements ModelInterface, ArrayAccess
         'volume_total' => 'int',
         'volume_remain' => 'int',
         'min_volume' => 'int',
-        'price' => 'float',
+        'price' => 'double',
         'is_buy_order' => 'bool',
         'duration' => 'int',
         'issued' => '\DateTime',
@@ -82,7 +82,7 @@ class GetMarketsRegionIdOrders200Ok implements ModelInterface, ArrayAccess
         'volume_total' => 'int32',
         'volume_remain' => 'int32',
         'min_volume' => 'int32',
-        'price' => 'float',
+        'price' => 'double',
         'is_buy_order' => null,
         'duration' => 'int32',
         'issued' => 'date-time',
@@ -525,7 +525,7 @@ class GetMarketsRegionIdOrders200Ok implements ModelInterface, ArrayAccess
     /**
      * Gets price
      *
-     * @return float
+     * @return double
      */
     public function getPrice()
     {
@@ -535,7 +535,7 @@ class GetMarketsRegionIdOrders200Ok implements ModelInterface, ArrayAccess
     /**
      * Sets price
      *
-     * @param float $price price number
+     * @param double $price price number
      *
      * @return $this
      */
